Update antiForgeryToken.js
There can be two `__RequestVerificationTokens` inputs, one from the _Host and one from the component `AntiForgeryTokenInput.razor` which is used manually in forms like on the `MainLayout.razor` for POSTing to /logout.

Some components/pages might try to obtain the token and get the input from `AntiForgeryTokenInput.razor` which has not yet been initialized, and will be null.

This simple fix sorts the inputs by the length of their value, thus putting the non-null ones at the top.